Rename ray.operator module
